Tweak addQueue skeleton; move meeting type code to separate module
